+++ src/components/CodeHighlight/index.ts
@@ -1,3 +1,3 @@
-// CodeHighlight from copy-ui @ 2025-09-12
+// CodeHighlight from copy-ui @ 2025-09-26
 
 export { CodeHighlight } from "./CodeHighlight";