+++ src/components/CodeHighlight/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-09-12: Initial version.
+- 2025-09-26: Add support for TSX language in CodeHighlight.